REFACTOR: widgetType class elements separated
