+++ src/renderer/src/cm-extensions/BlockQuoteExtension.ts
@@ -1,7 +1,6 @@
 import { syntaxTree } from '@codemirror/language';
-import { Decoration, EditorView, ViewPlugin, WidgetType } from '@codemirror/view';
+import { Decoration, ViewPlugin } from '@codemirror/view';
 import { Range } from '@codemirror/state';
-import * as emojiToolkit from 'emoji-toolkit';
 
 const tokenFormattingClasses = {
   'Blockquote': (isActive: boolean) => { return isActive ? 
@@ -50,57 +49,3 @@ export const BlockQuoteExtension = ViewPlugin.fromClass(class {
 }, {
   decorations: v => v.decorations
 });
-
-
-class EmojiWidget extends WidgetType {
-  emoji: string;
-
-  constructor(emoji: string) {
-    super();
-    this.emoji = emojiToolkit.shortnameToUnicode(emoji);  
-  }
-
-  toDOM(view: EditorView) {
-    const span = document.createElement('span');
-    span.textContent = this.emoji;
-    span.classList.add('cm-formatting-emoji');
-    return span;
-  }
-}
-
-class TaskMarkerWidget extends WidgetType {
-  isChecked: boolean;
-  position: { from: number, to: number };
-
-  constructor(isChecked: boolean, position: { from: number, to: number }) { 
-    super();
-    this.isChecked = isChecked;
-    this.position = position;
-  }
-
-  toDOM(view: EditorView) {
-    const input = document.createElement('input');
-    input.type = 'checkbox';
-    input.checked = this.isChecked;
-    input.classList.add('cm-formatting-input');
-
-    // event listener to update the state
-    input.addEventListener('change', (e) => {
-      this.toggle(view);
-    })
-
-    return input;
-  }
-
-  toggle(view: EditorView) {
-    const { from, to } = this.position;
-
-    const newContent = view.state.sliceDoc(from, to) === '[ ]' ? '[x]' : '[ ]';
-
-    view.dispatch({
-      changes: { from, to, insert: newContent },
-    });
-    view.focus();
-  }
-
-}
